requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
gradio>=4.0.0

# OpenAI and AI APIs
openai>=1.0.0
anthropic>=0.25.0

# Hugging Face
huggingface-hub>=0.20.0
transformers>=4.35.0

# Web Scraping
beautifulsoup4>=4.12.0
playwright>=1.40.0
lxml>=4.9.0

# Audio Processing
pydub>=0.25.0

# Image Processing
Pillow>=10.0.0

# Travel API
amadeus>=8.0.0

# Ollama (Local LLM)
ollama>=0.1.0

# Development and Utilities
jupyter>=1.0.0
ipython>=8.0.0
pandas>=2.0.0
numpy>=1.24.0 